AudioRecord

+++ app/src/main/java/com/lx/linwanandroid_mvvm/audioRecord/AudioBean.kt
@@ -7,4 +7,8 @@ package com.lx.linwanandroid_mvvm.audioRecord
  * @author linxiao
  * @data Created in 2021/05/01
  */
-class AudioBean(var fileName: String, var audioDuration: Long, var filePath: String, var isSelect: Boolean = false, var progress: Int = 0)
+class AudioBean(var fileName: String,
+                var audioDuration: Int,
+                var filePath: String,
+                var isSelect: Boolean = false,
+                var progress: Int = 0)
